extract ScatterplotExLayer to separate file
